Updates after move from `grails-core` (#6)
* chore: update project after move

Miscellaneous updates after project move
from `grails-core` to its own repo.

* chore: apply changes from feedback

* chore(ci): some more token cleanup

Clean up some more token params where it is already
set by default in the actions used.

* ci: add `GITHUB_TOKEN` to `actions/checkout`

This should not be needed when checking out the
"local" repository, but we have had problems
with `Input required and not supplied: token`
errors for this type of checkout.